requirements.txt
chinese_calendar
lxml
openai==0.27.6
pandas
pyyaml
requests
schedule
pyhandytools
sparkdesk-api==1.3.0
wcferry>=39.0.7.0
websocket
pillow
jupyter_client
zhdate
ipykernel
bardapi
